parsey: LLP64 cast, depend check non-gating in CRuby workflow
- The append-offset in pm_ystring_cat is ptrdiff_t (64-bit on LLP64
  Windows, where long is 32); cast explicitly as string lengths are
  long everywhere else.
- The CRuby bindings depend check reports its diff without failing:
  the parse.y backend's objects are patched into common.mk by the
  workflow, so their dependency lines are necessarily missing from
  CRuby's checked-in depend files until CRuby carries the change.
  (test-all itself is green on both parser matrices: 37778 tests,
  0 failures, with the backend linked in.)

+++ src/parsey/ystring.c
@@ -202,7 +202,7 @@ pm_ystring_cat(pm_ystring_t *str, const char *ptr, long len) {
      * the realloc below would move out from under them. Remember where they
      * are so they can be found again afterwards. */
     long off = -1;
-    if (ptr >= str->ptr && ptr <= str->ptr + str->len) off = ptr - str->ptr;
+    if (ptr >= str->ptr && ptr <= str->ptr + str->len) off = (long) (ptr - str->ptr);
 
     long total = str->len + len;
     char *reallocated = xrealloc_sized(str->ptr, (size_t) total + PM_YSTRING_TERM_LEN, (size_t) str->len + PM_YSTRING_TERM_LEN);
